+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 - Added managed and unmanaged exchange clients support to `ApiClient`.
 - Added `ApiClient.exchangeClient` to create exchange client builder.
 - Added `ApiClient.with` as an alias to `ApiClient.exchangeClient` to create exchange client builder.
-- Added `ExchangeClient.Builder` for managed exchange clients in `ApiClient`.
+- Added `ExchangeClientBuilder` for managed exchange clients in `ApiClient`.
 - Changed the default exchange client management to not managed when the `ApiClient` is built with direct `ExchangeClient` objects.
 - Added `ApiPage.of` to construct `ApiPage` objects.